Develop functions in the wbgapi_module.py module to receive data from wbgapi and write it to files.
Develop a function create_dfr that creates and transforms a pandas Data Frame.
Develop a write_dfr function that writes pandas Data Frame data to files.
Develop a function indicators_list that receives a list of statistics database indicators from wbgapi in the form of a table and writes it to a file.
Develop an auto_naming function that automatically assigns a name (in the form of the date and time at which the file was created) to new files that are written.